combine into commands/channel
